requirements.txt
python-docx>=1.1
pandas>=2.2
pyyaml>=6.0
